+++ src/DfE.ExternalApplications.Application/Templates/QueryObjects/GetCustomApplicationStatusesByTemplateIdQueryObject.cs
@@ -1,0 +1,23 @@
+using DfE.ExternalApplications.Application.Common.QueriesObjects;
+using DfE.ExternalApplications.Domain.Entities;
+using DfE.ExternalApplications.Domain.ValueObjects;
+using System;
+using System.Linq;
+
+namespace DfE.ExternalApplications.Application.Templates.QueryObjects
+{
+    public sealed class GetCustomApplicationStatusesByTemplateIdQueryObject : IQueryObject<CustomApplicationStatus>
+    {
+        private readonly TemplateId _templateId;
+
+        public GetCustomApplicationStatusesByTemplateIdQueryObject(Guid templateId)
+        {
+            _templateId = new TemplateId(templateId);
+        }
+
+        public IQueryable<CustomApplicationStatus> Apply(IQueryable<CustomApplicationStatus> query)
+        {
+            return query.Where(x => x.TemplateId == _templateId).OrderBy(x => x.ApplicationStatus);
+        }
+    }
+}